Move json util to common json
