+++ src/ai/backend/common/utils.py
@@ -20,14 +20,12 @@ from typing import (
     Iterator,
     Mapping,
     Optional,
-    Protocol,
     Tuple,
     TypeVar,
     Union,
 )
 
 import aiofiles
-import orjson
 import yarl
 from async_timeout import timeout
 
@@ -436,51 +434,3 @@ def b64encode(s: str) -> str:
     """
     b: bytes = s.encode("utf-8") if isinstance(s, str) else s
     return base64.b64encode(b).decode("ascii")
-
-
-def load_json(s: Union[bytes, bytearray, memoryview, str]) -> Any:
-    """
-    Loads a JSON string into a Python object.
-    """
-    return orjson.loads(s)
-
-
-class AsyncReader(Protocol):
-    async def read(self) -> bytes:
-        pass
-
-
-async def read_json(reader: AsyncReader) -> Any:
-    """
-    Reads a JSON body from an asynchronous reader.
-    """
-    data = await reader.read()
-    return load_json(data)
-
-
-def dump_json_str(obj: Any, option: Optional[int] = None) -> str:
-    """
-    Dumps the given object into a JSON string.
-    """
-    return orjson.dumps(obj, option=option).decode("utf-8")
-
-
-def dump_json(obj: Any, option: Optional[int] = None) -> bytes:
-    """
-    Dumps the given object into a JSON bytes.
-    """
-    return orjson.dumps(obj, option=option)
-
-
-def pretty_json_str(obj: Any) -> str:
-    """
-    Dumps the given object into a pretty JSON string.
-    """
-    return dump_json_str(obj, orjson.OPT_INDENT_2)
-
-
-def pretty_json(obj: Any) -> bytes:
-    """ "
-    Dumps the given object into a pretty JSON bytes.
-    """
-    return dump_json(obj, orjson.OPT_INDENT_2)
