+++ src/ai/backend/manager/api/cluster_template.py
@@ -10,7 +10,7 @@ import yaml
 from aiohttp import web
 
 from ai.backend.common import validators as tx
-from ai.backend.common.utils import load_json
+from ai.backend.common.json import load_json
 from ai.backend.logging import BraceStyleAdapter
 
 from ..models import (
