+++ src/ai/backend/manager/api/stream.py
@@ -45,8 +45,8 @@ from etcd_client import GRPCStatusCode, GRPCStatusError
 from ai.backend.common import redis_helper
 from ai.backend.common import validators as tx
 from ai.backend.common.events import KernelTerminatingEvent
+from ai.backend.common.json import dump_json, load_json
 from ai.backend.common.types import AccessKey, AgentId, KernelId, SessionId
-from ai.backend.common.utils import dump_json, load_json
 from ai.backend.logging import BraceStyleAdapter
 from ai.backend.manager.idle import AppStreamingStatus
