+++ src/ai/backend/manager/container_registry/docker.py
@@ -7,7 +7,7 @@ import typing_extensions
 import yarl
 
 from ai.backend.common.docker import login as registry_login
-from ai.backend.common.utils import read_json
+from ai.backend.common.json import read_json
 from ai.backend.logging import BraceStyleAdapter
 from ai.backend.manager.exceptions import ContainerRegistryProjectEmpty
