+++ src/ai/backend/manager/container_registry/local.py
@@ -9,7 +9,7 @@ import sqlalchemy as sa
 import yarl
 
 from ai.backend.common.docker import arch_name_aliases, get_docker_connector
-from ai.backend.common.utils import pretty_json_str
+from ai.backend.common.json import pretty_json_str
 from ai.backend.logging import BraceStyleAdapter
 
 from ..models.image import ImageRow, ImageStatus
