+++ src/ai/backend/storage/volumes/weka/weka_client.py
@@ -9,8 +9,8 @@ from typing import Any, Iterable, Mapping, MutableMapping, Optional
 import aiohttp
 from aiohttp import web
 
+from ai.backend.common.json import dump_json_str
 from ai.backend.common.types import BinarySize
-from ai.backend.common.utils import dump_json_str
 from ai.backend.logging import BraceStyleAdapter
 
 from .exceptions import WekaAPIError, WekaInvalidBodyError, WekaNotFoundError, WekaUnauthorizedError
